+++ internal/interpreter/zz_interpreter_flash_test.go
@@ -31,6 +31,8 @@ func (f *FakeVideoDeviceFlash) PrintChar(r rune)     {}
 func (f *FakeVideoDeviceFlash) PrintString(s string) {}
 func (f *FakeVideoDeviceFlash) SetCursorX(x int)     {}
 func (f *FakeVideoDeviceFlash) SetCursorY(y int)     {}
+func (f *FakeVideoDeviceFlash) CursorX() int         { return 0 }
+func (f *FakeVideoDeviceFlash) CursorY() int         { return 0 }
 func (f *FakeVideoDeviceFlash) SwitchMode(slot int)  {}
 func (f *FakeVideoDeviceFlash) Plot(x, y int)        {}
 func (f *FakeVideoDeviceFlash) ReadLine() (string, error) {