+++ internal/interpreter/zz_interpreter_val_assignment_test.go
@@ -17,6 +17,8 @@ func (m *MockVideo) PrintChar(r rune)     {}
 func (m *MockVideo) PrintString(s string) {}
 func (m *MockVideo) SetCursorX(x int)     {}
 func (m *MockVideo) SetCursorY(y int)     {}
+func (m *MockVideo) CursorX() int         { return 0 }
+func (m *MockVideo) CursorY() int         { return 0 }
 func (m *MockVideo) SwitchMode(slot int)  {}
 func (m *MockVideo) Plot(x, y int)        {}
 func (m *MockVideo) ReadLine() (string, error) {